requirements.txt
pyserial
numpy
rerun-sdk